clients: move Gatekeeper components
