+++ clients/apps/web/src/components/Gatekeeper/TakeoverBox.tsx
