+++ clients/apps/web/src/components/Gatekeeper/TakeoverHeader.tsx
